Do not add -L<dir> to the cm{x}a
This breaks on OSX when using libffi and ctypes, as ctypes requires
the version of libffi installed by brew, not the system one (installed
in /usr/lib`).

As we have no way to keep the order of linker flags (due to
https://caml.inria.fr/mantis/view.php?id=7150) it is safer to just not
use -L if possible. In the case of camlzip, this doesn't seem needed.